requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
teradatasql==17.20.0.31
cx_Oracle==8.3.0
snowflake-connector-python==3.6.0
pandas==2.1.4
python-decouple==3.8 